+++ glui/include/GL/glui.h
@@ -1526,24 +1526,6 @@ namespace glui {
     void            show( void );
     void            hide( void );
 
-    /***** GLUT callback setup functions *****/
-    /*
-      void set_glutDisplayFunc(void (*f)(void));
-      void set_glutReshapeFunc(void (*f)(int width, int height));
-      void set_glutKeyboardFunc(void (*f)(unsigned char key, int x, int y));
-      void set_glutSpecialFunc(void (*f)(int key, int x, int y));
-      void set_glutMouseFunc(void (*f)(int button, int state, int x, int y));
-      void set_glutMotionFunc(void (*f)(int x, int y));
-      void set_glutPassiveMotionFunc(void (*f)(int x, int y));
-      void set_glutEntryFunc(void (*f)(int state));
-      void set_glutVisibilityFunc(void (*f)(int state));
-      void set_glutInit( int *argcp, const char **argv );
-      void set_glutInitWindowSize(int width, int height);
-      void set_glutInitWindowPosition(int x, int y);
-      void set_glutInitDisplayMode(unsigned int mode);
-      int  set_glutCreateWindow(const char *name);
-    */
-
     /***** Constructors and desctructors *****/
 
     int init( const char *name, long flags, int x, int y, int parent_window );
@@ -2463,85 +2445,6 @@ namespace glui {
     void common_init();
   };
 
-  /************************************************************/
-  /*                                                          */
-  /*                   Translation class                      */
-  /*                                                          */
-  /************************************************************/
-
-  /**
-     An onscreen translation controller--allows the user to interact with
-     a 3D translation.
-  */
-  class GLUI_API GLUI_Translation : public GLUI_Mouse_Interaction
-  {
-  public:
-    int trans_type;  /* Is this an XY or a Z controller? */
-    int down_x, down_y;
-    float scale_factor;
-    GLUquadricObj *quadObj;
-    int   trans_mouse_code;
-    float orig_x, orig_y, orig_z;
-    int   locked;
-
-    int  iaction_mouse_down_handler( int local_x, int local_y );
-    int  iaction_mouse_up_handler( int local_x, int local_y, bool inside );
-    int  iaction_mouse_held_down_handler( int local_x, int local_y, bool inside );
-    int  iaction_special_handler( int key, int modifiers );
-    void iaction_init( void ) { }
-    void iaction_draw_active_area_persp( void );
-    void iaction_draw_active_area_ortho( void );
-    void iaction_dump( FILE *output );
-
-    void set_speed( float s ) { scale_factor = s; }
-
-    void setup_texture( void );
-    void setup_lights( void );
-    void draw_2d_arrow( int radius, int filled, int orientation ); 
-    void draw_2d_x_arrows( int radius );
-    void draw_2d_y_arrows( int radius );
-    void draw_2d_z_arrows( int radius );
-    void draw_2d_xy_arrows( int radius );
-
-    int  get_mouse_code( int x, int y );
-
-    /* Float array is either a single float (for single-axis controls),
-       or two floats for X and Y (if an XY controller) */
-
-    float get_z( void ) {       return float_array_val[0];  }
-    float get_x( void ) {       return float_array_val[0];  }
-    float get_y( void ) {
-      if ( trans_type == GLUI_TRANSLATION_XY )    return float_array_val[1];
-      else					return float_array_val[0];
-    }
-
-    void  set_z( float val );
-    void  set_x( float val );
-    void  set_y( float val );
-    void  set_one_val( float val, int index );
-
-    GLUI_Translation( GLUI_Node *parent, const char *name,
-                      int trans_type, float *live_var=NULL,
-                      int id=-1, CallBack callback=CallBack()	);
-    GLUI_Translation( void ) { common_init(); }
-
-  protected:
-    void common_init() {
-      locked              = GLUI_TRANSLATION_LOCK_NONE;
-      glui_format_str( name, "Translation: %p", this );
-      w                   = GLUI_MOUSE_INTERACTION_WIDTH;
-      h                   = GLUI_MOUSE_INTERACTION_HEIGHT;
-      can_activate        = true;
-      live_type           = GLUI_LIVE_FLOAT_ARRAY;
-      float_array_size    = 0;
-      alignment           = GLUI_ALIGN_CENTER;
-      trans_type          = GLUI_TRANSLATION_XY;
-      scale_factor        = 1.0;
-      quadObj             = NULL;
-      trans_mouse_code    = GLUI_TRANSLATION_MOUSE_NONE;
-    }
-  };
-
   /********** Misc functions *********************/
   int _glutBitmapWidthString( void *font, const char *s );
   void _glutBitmapString( void *font, const char *s );
@@ -2569,3 +2472,6 @@ namespace glui {
   // void glui_parent_window_special_func(int key, int x, int y);
 
 }
+
+
+#include "glui/Translation.h"
